Revert the change in composer file.
They changed the notation for packages.drupal.org that is now the official service.

"When using the official packages.drupal.org service, the platform part of the version
is already defined by whether you are using the Drupal 7 or Drupal 8 endpoint,
therefore the rest of the translated format is as follows: Major.Minor.0

e.g. drupal/views": "3.13.*"

More Details here:
https://www.drupal.org/docs/develop/using-composer/using-packagesdrupalorg