Issue #2934102 by skyredwang, olafkarsten: Implement a PriceResolver
Adds BundlePriceResolver class.
Extends the BundleInterface with setPrice() method.
Refactored the ProductBundle getPrice() method. It now returns NULL if no price is set on the object. The logic for calculating a derived bundle price from the bundle items/referenced variations has moved to the price resolver.
Adds tests for all that stuff.